refactor: 重構 AuthorInfo 和 ContactUs 的資料夾位置，以方便管理。
