+++ src/pages/contact-us/index.tsx
@@ -1,5 +1,5 @@
 import { NextPage } from 'next';
-import ContactUs from '@/components/ContactUs';
+import ContactUs from '@/components/pages/ContactUs';
 
 const ContactUsPage: NextPage = () => {
   return <ContactUs />;
